requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
django-tabbed-admin==1.0.4
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
